refactor: dedup codebase
