+++ src/v2/createEmailTemplate.ts
@@ -1,8 +1,6 @@
 import type {RequestHandler} from 'express';
-import {
-	hasTemplate, setTemplate, templateSchema,
-} from '../store';
-import {compileTemplateParts, TemplateRenderError} from './renderTemplate';
+import {hasTemplate, setTemplate, templateSchema} from '../store';
+import {compileTemplateOrReject} from './templateRequest';
 
 const handler: RequestHandler = (req, res) => {
 	const result = templateSchema.safeParse({...req.body, CreatedTimestamp: Date.now() / 1000});
@@ -19,14 +17,7 @@ const handler: RequestHandler = (req, res) => {
 
 	// Reject malformed Handlebars up front, as real SES does at create time (rather than
 	// hard-failing every later send). Compilation catches only syntax errors, not missing variables.
-	try {
-		compileTemplateParts(result.data.TemplateContent);
-	} catch (error: unknown) {
-		if (!(error instanceof TemplateRenderError)) {
-			throw error;
-		}
-
-		res.status(400).send({type: 'BadRequestException', message: 'Bad Request Exception', detail: `aws-ses-v2-local: Handlebars compilation failed - ${error.message}`});
+	if (!compileTemplateOrReject(result.data.TemplateContent, res)) {
 		return;
 	}
